# apps/backend/openapi/platform-api.json
{
  "components": {
...
    "securitySchemes": {
      "logto": {
        "bearerFormat": "JWT",
        "description": "Logto access token for the exact Platform API audience.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Canonical REST contract for the Inside Platform web and agent adapters.",
    "title": "Inside Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/library/materials": {
      "get": {
        "operationId": "listPublishedMaterials",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "items": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "access": {
                            "enum": [
                              "free",
                              "membership"
                            ],
                            "type": "string"
                          },
                          "availability": {
                            "enum": [
                              "available",
                              "locked",
                              "unavailable"
                            ],
                            "type": "string"
                          },
                          "contentVersion": {
                            "exclusiveMinimum": true,
                            "maximum": 9007199254740991,
                            "minimum": 0,
                            "type": "integer"
                          },
                          "format": {
                            "additionalProperties": false,
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "slug"
                            ],
                            "type": "object"
                          },
                          "materialId": {
                            "type": "string"
                          },
                          "publishedAt": {
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                            "type": "string"
                          },
                          "seriesMemberships": {
                            "items": {
                              "additionalProperties": false,
                              "properties": {
                                "ordinal": {
                                  "exclusiveMinimum": true,
                                  "maximum": 9007199254740991,
                                  "minimum": 0,
                                  "type": "integer"
                                },
                                "series": {
                                  "additionalProperties": false,
                                  "properties": {
                                    "id": {
                                      "type": "string"
                                    },
                                    "name": {
                                      "type": "string"
                                    },
                                    "slug": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "name",
                                    "slug"
                                  ],
                                  "type": "object"
                                }
                              },
                              "required": [
                                "ordinal",
                                "series"
                              ],
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "summary": {
                            "type": "string"
                          },
                          "tags": {
                            "items": {
                              "additionalProperties": false,
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "id",
                                "name"
                              ],
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "title": {
                            "type": "string"
                          },
                          "topic": {
                            "additionalProperties": false,
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "slug"
                            ],
                            "type": "object"
                          }
                        },
                        "required": [
                          "materialId",
                          "contentVersion",
                          "slug",
                          "title",
                          "summary",
                          "access",
                          "publishedAt",
                          "topic",
                          "format",
                          "tags",
                          "seriesMemberships",
                          "availability"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "maxLength": 512,
                      "minLength": 1,
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "A deterministic page of published Materials"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "correlationId": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "status": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Catalog cursor is malformed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "correlationId": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "status": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Optional Account proof is invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "status": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "title": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "title",
                        "status",
                        "code"
                      ],
                      "type": "object"
                    },
                    {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "detail": {
                          "type": "string"
                        },
                        "status": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "title": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "title",
                        "status",
                        "detail",
                        "code"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "Catalog or Account resolution failed internally"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "status": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "title": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "title",
                        "status",
                        "code"
                      ],
                      "type": "object"
                    },
                    {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "detail": {
                          "type": "string"
                        },
                        "status": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "title": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "title",
                        "status",
                        "detail",
                        "code"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "Catalog or Account proof dependency is unavailable"
          }
        },
        "security": [
          {},
          {
            "logto": []
          }
        ],
        "summary": "List safe published Material projections",
        "tags": [
          "Content library"
        ]
      }
    },
...
  },
  "servers": [],
...
}